pkg/ioutilx: move Windows path helpers to pkg/fsutil
WindowsSubsystemPath, WindowsSubsystemPathWithCygpath,
WindowsSubsystemPathForLinux, and the unexported
windowsSubsystemPathWithoutCygpath fallback do not resemble io/ioutil
and belong with the other filesystem helpers in pkg/fsutil, per #5283.
ReadAtMaximum and the UTF16 helpers stay in pkg/ioutilx; their eventual
home is still an open question in the issue.

The real implementation moves into pkg/fsutil/fsutil_windows.go, which
is only compiled on GOOS=windows per its filename suffix. All existing
callers (cmd/limactl/shell.go, pkg/copytool, pkg/hostagent,
pkg/sshutil, pkg/limayaml/defaults.go) are cross-platform files that
already gate these calls behind "runtime.GOOS == \"windows\""
checks, so the symbols still need to resolve at compile time on
non-Windows. pkg/fsutil/fsutil_unix.go (//go:build !windows) adds
stub definitions with matching signatures that return an error; they
are unreachable in practice since every call site is
runtime.GOOS-gated, mirroring the existing errNotSupported pattern in
pkg/guestagent/timesync/timesync_others.go. This keeps every caller's
code unchanged and keeps the diff minimal while still not breaking
any GOOS build.

The Windows-only unit test (pkg/ioutilx/ioutilx_windows_test.go) moves
alongside the code to pkg/fsutil/fsutil_windows_test.go, unchanged
apart from the package name.

No compatibility alias/stub was kept in pkg/ioutilx for the moved
symbols: recent pkg/ moves in this repo's history (e.g. the
sshutil/ioutilx cygpath work in 7a5a207 and 54db504) rewrite call
sites directly rather than leaving forwarding shims, so that
convention is followed here too.

Verified: go build ./... on darwin, GOOS=linux go build ./...,
GOOS=windows go build ./... (all clean); go test
./pkg/ioutilx/... ./pkg/fsutil/... plus the caller packages
(copytool, hostagent, sshutil, limayaml, cmd/limactl) on darwin;
GOOS=windows go vet/test -c on pkg/fsutil to compile-check the
Windows-only files; go vet ./... on darwin; golangci-lint run on all
touched packages (0 issues).
