+++ pkg/fsutil/fsutil_windows_test.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright The Lima Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package ioutilx
+package fsutil
 
 import (
 	"strings"
